✨ feat: Add autoplay video in HTML with Rick Astley's "Never Gonna Give You Up"
The commit includes the addition of a new HTML file that autoplays a video of Rick Astley's song "Never Gonna Give You Up" when the webpage is loaded. The video plays in a loop with sound at maximum volume. The webpage is styled to be full screen with no overflow and a black background.